feat: add event encoding and decoding
There were no explicit encoding and decoding of events for the network,
which means that some data was unnecessarily large and could for that
reason cause messages to exceed the MTU. To partially reduce the risk,
this commit adds a module to handle encoding and decoding of events in a
more dynamic format. It also guarantees that the format stays
well-defined, even if fields are added and removed.

+++ swim/event.h
@@ -96,7 +96,7 @@ typedef struct Event {
     struct LeaveEvent leave;
   };
 
-  int gossip_count;
+  uint16_t gossip_count;
   NodeInfo gossip[];
 } Event;
